refactor: Структура
